pyproject.toml
[tool.poetry]
name = "thumbnails"
version = "0.1.0"
description = ""
authors = ["yukinakanaka <yuki.nakamura@mapbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.28.30"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
taskipy = "^1.12.0"
autopep8 = "^2.0.2"
pytest = "^7.4.0"
types-pillow = "^10.0.0.2"
boto3-stubs = { extras = ["s3"], version = "^1.28.30" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests", "integration"]

[tool.taskipy.tasks]
lint = { cmd = "task lint_black && task lint_flake8 && task lint_mypy", help = "exec lint" }
lint_black = "black --check src/"
lint_flake8 = "flake8 --max-line-length 169 src/"
lint_mypy = "mypy src"

fix_lint = { cmd = "autopep8 -i --aggressive --max-line-length 169 --recursive src/" }
fix_format = { cmd = "black src/ tests/ && isort src/ tests/", help = "format code" }
